+++ src/renderable/imagelayer.js
@@ -235,11 +235,17 @@ import * as stringUtil from "./../utils/string.js";
     preDraw(renderer) {
         // save the context
         renderer.save();
+
         // apply the defined alpha value
         renderer.setGlobalAlpha(renderer.globalAlpha() * this.getOpacity());
 
         // apply the defined tint, if any
         renderer.setTint(this.tint);
+
+        // apply blending if different from "normal"
+        if (this.blendMode !== renderer.getBlendMode()) {
+            renderer.setBlendMode(this.blendMode);
+        }
     }
 
     /**